Move the trio code into its own file so that it can be excluded on Python < 3.5
